UINV-57 refactoring presentation componet
