+++ lib/FundDistribution/FundDistributionFields/FundDistributionFieldsContainer.js
@@ -1,18 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {
-  FieldArray,
-} from 'redux-form';
-import { FormattedMessage } from 'react-intl';
-import { get } from 'lodash';
 
 import {
   stripesConnect,
 } from '@folio/stripes/core';
-
-import {
-  RepeatableField,
-} from '@folio/stripes/components';
 
 import { getFundsForSelect } from '../../utils';
 import { baseManifest } from '../../manifests';
@@ -23,35 +14,14 @@ const FundDistributionFieldsContainer = ({
 }) => {
   const funds = getFundsForSelect(resources);
 
-  const renderSubForm = (elem, index, fields) => {
-    const fund = fields.get(index);
-    const fundCode = get(funds.find(({ value }) => value === fund.fundId), 'code');
-
-    return (
-      <FundDistributionFields
-        fundCode={fundCode}
-        funds={funds}
-        fund={fund}
-        totalAmount={totalAmount}
-        required={required}
-        elem={elem}
-        disabled={disabled}
-        currency={currency}
-      />
-    );
-  };
-
   return (
-    <FieldArray
-      addLabel={<FormattedMessage id="stripes-acq-components.fundDistribution.addBtn" />}
-      component={RepeatableField}
-      name="fundDistributions"
-      props={{
-        canAdd: !disabled,
-        canRemove: !disabled,
-        formValues,
-      }}
-      renderField={renderSubForm}
+    <FundDistributionFields
+      formValues={formValues}
+      funds={funds}
+      totalAmount={totalAmount}
+      required={required}
+      disabled={disabled}
+      currency={currency}
     />
   );
 };
